Move core integrations from extended to baseline
