+++ test/e2e/singlecluster/extended/tas_test.go
@@ -27,16 +27,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta2"
-	workloadjob "sigs.k8s.io/kueue/pkg/controller/jobs/job"
 	"sigs.k8s.io/kueue/pkg/controller/jobs/jobset"
-	podcontroller "sigs.k8s.io/kueue/pkg/controller/jobs/pod"
-	podconstants "sigs.k8s.io/kueue/pkg/controller/jobs/pod/constants"
 	workloadtrainjob "sigs.k8s.io/kueue/pkg/controller/jobs/trainjob"
 	"sigs.k8s.io/kueue/pkg/util/tas"
 	utiltestingapi "sigs.k8s.io/kueue/pkg/util/testing/v1beta2"
-	testingjob "sigs.k8s.io/kueue/pkg/util/testingjobs/job"
 	testingjobset "sigs.k8s.io/kueue/pkg/util/testingjobs/jobset"
-	testingpod "sigs.k8s.io/kueue/pkg/util/testingjobs/pod"
 	testingtrainjob "sigs.k8s.io/kueue/pkg/util/testingjobs/trainjob"
 	"sigs.k8s.io/kueue/test/util"
 )
@@ -50,94 +45,6 @@ var _ = ginkgo.Describe("TopologyAwareScheduling", ginkgo.Label("area:singleclus
 	ginkgo.AfterEach(func() {
 		gomega.Expect(util.DeleteNamespace(ctx, k8sClient, ns)).To(gomega.Succeed())
 		util.ExpectAllPodsInNamespaceDeleted(ctx, k8sClient, ns)
-	})
-
-	ginkgo.When("Creating a Job requesting TAS", func() {
-		var (
-			topology     *kueue.Topology
-			onDemandRF   *kueue.ResourceFlavor
-			localQueue   *kueue.LocalQueue
-			clusterQueue *kueue.ClusterQueue
-		)
-		ginkgo.BeforeEach(func() {
-			topology = utiltestingapi.MakeDefaultOneLevelTopology("hostname-" + ns.Name)
-			util.MustCreate(ctx, k8sClient, topology)
-
-			onDemandRF = utiltestingapi.MakeResourceFlavor("on-demand-"+ns.Name).
-				NodeLabel("instance-type", "on-demand").TopologyName(topology.Name).Obj()
-			util.MustCreate(ctx, k8sClient, onDemandRF)
-			clusterQueue = utiltestingapi.MakeClusterQueue("cluster-queue-" + ns.Name).
-				ResourceGroup(
-					*utiltestingapi.MakeFlavorQuotas(onDemandRF.Name).
-						Resource(corev1.ResourceCPU, "1").
-						Resource(corev1.ResourceMemory, "1Gi").
-						Obj(),
-				).
-				Obj()
-			util.CreateClusterQueuesAndWaitForActive(ctx, k8sClient, clusterQueue)
-
-			localQueue = utiltestingapi.MakeLocalQueue("main", ns.Name).ClusterQueue(clusterQueue.Name).Obj()
-			util.CreateLocalQueuesAndWaitForActive(ctx, k8sClient, localQueue)
-		})
-		ginkgo.AfterEach(func() {
-			gomega.Expect(util.DeleteAllJobsInNamespace(ctx, k8sClient, ns)).Should(gomega.Succeed())
-			// Force remove workloads to be sure that cluster queue can be removed.
-			gomega.Expect(util.DeleteWorkloadsInNamespace(ctx, k8sClient, ns)).Should(gomega.Succeed())
-			gomega.Expect(util.DeleteObject(ctx, k8sClient, localQueue)).Should(gomega.Succeed())
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, clusterQueue, true)
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, onDemandRF, true)
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, topology, true)
-		})
-
-		ginkgo.It("should admit a Job via TAS", func() {
-			sampleJob := testingjob.MakeJob("test-job", ns.Name).
-				Queue(kueue.LocalQueueName(localQueue.Name)).
-				RequestAndLimit(corev1.ResourceCPU, "700m").
-				RequestAndLimit(corev1.ResourceMemory, "20Mi").
-				PodAnnotation(kueue.PodSetRequiredTopologyAnnotation, corev1.LabelHostname).
-				Image(util.GetAgnHostImage(), util.BehaviorExitFast).
-				Obj()
-			util.MustCreate(ctx, k8sClient, sampleJob)
-
-			createdWorkload := &kueue.Workload{}
-			wlLookupKey := types.NamespacedName{Name: workloadjob.GetWorkloadNameForJob(sampleJob.Name, sampleJob.UID), Namespace: ns.Name}
-
-			ginkgo.By(fmt.Sprintf("await for admission of workload %q", wlLookupKey), func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sClient.Get(ctx, wlLookupKey, createdWorkload)).Should(gomega.Succeed())
-					g.Expect(createdWorkload.Status.Admission).ShouldNot(gomega.BeNil())
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			// The job might have finished at this point. That shouldn't be a problem for the purpose of this test
-			jobKey := client.ObjectKeyFromObject(sampleJob)
-			util.ExpectJobUnsuspendedWithNodeSelectors(ctx, k8sClient, jobKey, map[string]string{
-				"instance-type": "on-demand",
-			})
-
-			ginkgo.By("verify TopologyAssignment", func() {
-				gomega.Expect(createdWorkload.Status.Admission.PodSetAssignments).Should(gomega.HaveLen(1))
-				gomega.Expect(createdWorkload.Status.Admission.PodSetAssignments[0].TopologyAssignment).Should(gomega.BeComparableTo(
-					tas.V1Beta2From(&tas.TopologyAssignment{
-						Levels: []string{
-							corev1.LabelHostname,
-						},
-						Domains: []tas.TopologyDomainAssignment{
-							{
-								Count: 1,
-								Values: []string{
-									"kind-worker",
-								},
-							},
-						},
-					}),
-				))
-			})
-
-			ginkgo.By(fmt.Sprintf("verify the workload %q gets finished", wlLookupKey), func() {
-				util.ExpectWorkloadToFinishWithTimeout(ctx, k8sClient, wlLookupKey, util.LongTimeout)
-			})
-		})
 	})
 
 	ginkgo.When("Creating a JobSet requesting TAS", func() {
@@ -269,161 +176,6 @@ var _ = ginkgo.Describe("TopologyAwareScheduling", ginkgo.Label("area:singleclus
 		})
 	})
 
-	ginkgo.When("Creating a Pod requesting TAS", func() {
-		var (
-			topology     *kueue.Topology
-			onDemandRF   *kueue.ResourceFlavor
-			clusterQueue *kueue.ClusterQueue
-			localQueue   *kueue.LocalQueue
-		)
-		ginkgo.BeforeEach(func() {
-			topology = utiltestingapi.MakeDefaultOneLevelTopology("hostname-" + ns.Name)
-			util.MustCreate(ctx, k8sClient, topology)
-
-			onDemandRF = utiltestingapi.MakeResourceFlavor("on-demand-"+ns.Name).
-				NodeLabel("instance-type", "on-demand").
-				TopologyName(topology.Name).
-				Obj()
-
-			util.MustCreate(ctx, k8sClient, onDemandRF)
-			clusterQueue = utiltestingapi.MakeClusterQueue("cluster-queue-" + ns.Name).
-				ResourceGroup(
-					*utiltestingapi.MakeFlavorQuotas(onDemandRF.Name).
-						Resource(corev1.ResourceCPU, "1").
-						Resource(corev1.ResourceMemory, "1Gi").
-						Obj(),
-				).
-				Obj()
-			util.CreateClusterQueuesAndWaitForActive(ctx, k8sClient, clusterQueue)
-
-			localQueue = utiltestingapi.MakeLocalQueue("main", ns.Name).ClusterQueue(clusterQueue.Name).Obj()
-			util.CreateLocalQueuesAndWaitForActive(ctx, k8sClient, localQueue)
-		})
-		ginkgo.AfterEach(func() {
-			gomega.Expect(util.DeleteAllJobsInNamespace(ctx, k8sClient, ns)).Should(gomega.Succeed())
-			// Force remove workloads to be sure that cluster queue can be removed.
-			gomega.Expect(util.DeleteWorkloadsInNamespace(ctx, k8sClient, ns)).Should(gomega.Succeed())
-			gomega.Expect(util.DeleteObject(ctx, k8sClient, localQueue)).Should(gomega.Succeed())
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, clusterQueue, true)
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, onDemandRF, true)
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, topology, true)
-		})
-
-		ginkgo.It("should admit a single Pod via TAS", func() {
-			p := testingpod.MakePod("test-pod", ns.Name).
-				Queue(localQueue.Name).
-				Image(util.GetAgnHostImage(), util.BehaviorExitFast).
-				Annotation(kueue.PodSetRequiredTopologyAnnotation, corev1.LabelHostname).
-				RequestAndLimit(corev1.ResourceCPU, "200m").
-				RequestAndLimit(corev1.ResourceMemory, "200Mi").
-				Obj()
-
-			ginkgo.By("Creating the Pod", func() {
-				util.MustCreate(ctx, k8sClient, p)
-				gomega.Expect(p.Spec.SchedulingGates).To(gomega.ContainElements(
-					corev1.PodSchedulingGate{Name: podconstants.SchedulingGateName},
-					corev1.PodSchedulingGate{Name: kueue.TopologySchedulingGate},
-				))
-			})
-
-			ginkgo.By("waiting for the Pod to be unsuspended", func() {
-				jobSetKey := client.ObjectKeyFromObject(p)
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sClient.Get(ctx, jobSetKey, p)).To(gomega.Succeed())
-					g.Expect(p.Spec.NodeSelector).To(gomega.Equal(map[string]string{
-						"instance-type":      "on-demand",
-						corev1.LabelHostname: "kind-worker",
-					}))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			wlLookupKey := types.NamespacedName{Name: podcontroller.GetWorkloadNameForPod(p.Name, p.UID), Namespace: ns.Name}
-			createdWorkload := &kueue.Workload{}
-
-			ginkgo.By(fmt.Sprintf("await for admission of workload %q and verify TopologyAssignment", wlLookupKey), func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sClient.Get(ctx, wlLookupKey, createdWorkload)).Should(gomega.Succeed())
-					g.Expect(createdWorkload.Status.Admission).ShouldNot(gomega.BeNil())
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-				gomega.Expect(createdWorkload.Status.Admission).ShouldNot(gomega.BeNil())
-				gomega.Expect(createdWorkload.Status.Admission.PodSetAssignments).Should(gomega.HaveLen(1))
-				gomega.Expect(createdWorkload.Status.Admission.PodSetAssignments[0].TopologyAssignment).Should(gomega.BeComparableTo(
-					tas.V1Beta2From(&tas.TopologyAssignment{
-						Levels: []string{corev1.LabelHostname},
-						Domains: []tas.TopologyDomainAssignment{{
-							Count:  1,
-							Values: []string{"kind-worker"},
-						}},
-					}),
-				))
-			})
-
-			ginkgo.By(fmt.Sprintf("verify the workload %q gets finished", wlLookupKey), func() {
-				util.ExpectWorkloadToFinishWithTimeout(ctx, k8sClient, wlLookupKey, util.LongTimeout)
-			})
-		})
-
-		ginkgo.It("should admit a Pod group via TAS", func() {
-			group := testingpod.MakePod("group", ns.Name).
-				Queue(localQueue.Name).
-				Image(util.GetAgnHostImage(), util.BehaviorExitFast).
-				Annotation(kueue.PodSetRequiredTopologyAnnotation, corev1.LabelHostname).
-				RequestAndLimit(corev1.ResourceCPU, "200m").
-				RequestAndLimit(corev1.ResourceMemory, "200Mi").
-				MakeGroup(2)
-
-			ginkgo.By("Creating the Pod group", func() {
-				for _, p := range group {
-					util.MustCreate(ctx, k8sClient, p)
-					gomega.Expect(p.Spec.SchedulingGates).To(gomega.ContainElements(
-						corev1.PodSchedulingGate{Name: podconstants.SchedulingGateName},
-						corev1.PodSchedulingGate{Name: kueue.TopologySchedulingGate},
-					))
-				}
-			})
-
-			ginkgo.By("waiting for the Pod to be ungated", func() {
-				// Verify that the Pods start with the appropriate selector.
-				gomega.Eventually(func(g gomega.Gomega) {
-					for _, origPod := range group {
-						var p corev1.Pod
-						g.Expect(k8sClient.Get(ctx, client.ObjectKeyFromObject(origPod), &p)).To(gomega.Succeed())
-						g.Expect(p.Spec.SchedulingGates).To(gomega.BeEmpty())
-						g.Expect(p.Spec.NodeSelector).To(gomega.Equal(map[string]string{
-							"instance-type":      "on-demand",
-							corev1.LabelHostname: "kind-worker",
-						}))
-					}
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			wlLookupKey := client.ObjectKey{Namespace: ns.Name, Name: "group"}
-			createdWorkload := &kueue.Workload{}
-
-			ginkgo.By(fmt.Sprintf("await for admission of workload %q and verify TopologyAssignment", wlLookupKey), func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sClient.Get(ctx, wlLookupKey, createdWorkload)).Should(gomega.Succeed())
-					g.Expect(createdWorkload.Status.Admission).ShouldNot(gomega.BeNil())
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-				gomega.Expect(createdWorkload.Status.Admission).ShouldNot(gomega.BeNil())
-				gomega.Expect(createdWorkload.Status.Admission.PodSetAssignments).Should(gomega.HaveLen(1))
-				gomega.Expect(createdWorkload.Status.Admission.PodSetAssignments[0].TopologyAssignment).Should(gomega.BeComparableTo(
-					tas.V1Beta2From(&tas.TopologyAssignment{
-						Levels: []string{corev1.LabelHostname},
-						Domains: []tas.TopologyDomainAssignment{{
-							Count:  2,
-							Values: []string{"kind-worker"},
-						}},
-					}),
-				))
-			})
-
-			ginkgo.By(fmt.Sprintf("verify the workload %q gets finished", wlLookupKey), func() {
-				util.ExpectWorkloadToFinishWithTimeout(ctx, k8sClient, wlLookupKey, util.LongTimeout)
-			})
-		})
-	})
-
 	ginkgo.When("Creating a TrainJob requesting TAS", func() {
 		var (
 			topology     *kueue.Topology
@@ -531,115 +283,6 @@ var _ = ginkgo.Describe("TopologyAwareScheduling", ginkgo.Label("area:singleclus
 
 			ginkgo.By(fmt.Sprintf("verify the workload %q gets finished", wlLookupKey), func() {
 				util.ExpectWorkloadToFinishWithTimeout(ctx, k8sClient, wlLookupKey, util.LongTimeout)
-			})
-		})
-	})
-
-	ginkgo.When("Scaling a Job requesting TAS via workload slices", func() {
-		var (
-			topology     *kueue.Topology
-			onDemandRF   *kueue.ResourceFlavor
-			localQueue   *kueue.LocalQueue
-			clusterQueue *kueue.ClusterQueue
-		)
-
-		ginkgo.BeforeEach(func() {
-			topology = utiltestingapi.MakeDefaultOneLevelTopology("hostname-" + ns.Name)
-			util.MustCreate(ctx, k8sClient, topology)
-
-			onDemandRF = utiltestingapi.MakeResourceFlavor("on-demand-"+ns.Name).
-				NodeLabel("instance-type", "on-demand").TopologyName(topology.Name).Obj()
-			util.MustCreate(ctx, k8sClient, onDemandRF)
-			clusterQueue = utiltestingapi.MakeClusterQueue("cluster-queue-" + ns.Name).
-				ResourceGroup(
-					*utiltestingapi.MakeFlavorQuotas(onDemandRF.Name).
-						Resource(corev1.ResourceCPU, "4").
-						Resource(corev1.ResourceMemory, "4Gi").
-						Obj(),
-				).
-				Obj()
-			util.CreateClusterQueuesAndWaitForActive(ctx, k8sClient, clusterQueue)
-
-			localQueue = utiltestingapi.MakeLocalQueue("main", ns.Name).ClusterQueue(clusterQueue.Name).Obj()
-			util.CreateLocalQueuesAndWaitForActive(ctx, k8sClient, localQueue)
-		})
-
-		ginkgo.AfterEach(func() {
-			gomega.Expect(util.DeleteAllJobsInNamespace(ctx, k8sClient, ns)).Should(gomega.Succeed())
-			// Force remove workloads to be sure that cluster queue can be removed.
-			gomega.Expect(util.DeleteWorkloadsInNamespace(ctx, k8sClient, ns)).Should(gomega.Succeed())
-			gomega.Expect(util.DeleteObject(ctx, k8sClient, localQueue)).Should(gomega.Succeed())
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, clusterQueue, true)
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, onDemandRF, true)
-			util.ExpectObjectToBeDeleted(ctx, k8sClient, topology, true)
-		})
-
-		ginkgo.It("should preserve topology assignment during scale-up", func() {
-			sampleJob := testingjob.MakeJob("test-job", ns.Name).
-				Queue(kueue.LocalQueueName(localQueue.Name)).
-				SetAnnotation("kueue.x-k8s.io/elastic-job", "true").
-				Parallelism(1).
-				Completions(10).
-				RequestAndLimit(corev1.ResourceCPU, "200m").
-				RequestAndLimit(corev1.ResourceMemory, "20Mi").
-				PodAnnotation(kueue.PodSetUnconstrainedTopologyAnnotation, "true").
-				Image(util.GetAgnHostImage(), util.BehaviorWaitForDeletion).
-				TerminationGracePeriod(1).
-				Obj()
-			util.MustCreate(ctx, k8sClient, sampleJob)
-
-			var createdWorkload *kueue.Workload
-			var originalNodeName string
-			ginkgo.By("await for admission of workload and record topology", func() {
-				// Use ExpectWorkloadsInNamespace instead of computing workload name
-				// because with workload slicing enabled, the name includes generation
-				workloads := util.ExpectWorkloadsInNamespace(ctx, k8sClient, ns.Name, 1)
-				createdWorkload = &workloads[0]
-
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sClient.Get(ctx, client.ObjectKeyFromObject(createdWorkload), createdWorkload)).Should(gomega.Succeed())
-					g.Expect(createdWorkload.Status.Admission).ShouldNot(gomega.BeNil())
-					g.Expect(createdWorkload.Status.Admission.PodSetAssignments).Should(gomega.HaveLen(1))
-					g.Expect(createdWorkload.Status.Admission.PodSetAssignments[0].TopologyAssignment).ShouldNot(gomega.BeNil())
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-
-				ta := tas.InternalFrom(createdWorkload.Status.Admission.PodSetAssignments[0].TopologyAssignment)
-				gomega.Expect(ta.Domains).ShouldNot(gomega.BeEmpty())
-				originalNodeName = ta.Domains[0].Values[0]
-			})
-
-			scaledParallelism := int32(2)
-			ginkgo.By("scale up the job parallelism", func() {
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sClient.Get(ctx, client.ObjectKeyFromObject(sampleJob), sampleJob)).To(gomega.Succeed())
-					sampleJob.Spec.Parallelism = new(scaledParallelism)
-					g.Expect(k8sClient.Update(ctx, sampleJob)).To(gomega.Succeed())
-				}, util.Timeout, util.Interval).Should(gomega.Succeed())
-			})
-
-			ginkgo.By("verify scaled workload preserves original topology assignment", func() {
-				scaledWorkload := util.ExpectNewWorkloadSlice(ctx, k8sClient, createdWorkload)
-				gomega.Expect(scaledWorkload).ShouldNot(gomega.BeNil())
-
-				gomega.Eventually(func(g gomega.Gomega) {
-					g.Expect(k8sClient.Get(ctx, client.ObjectKeyFromObject(scaledWorkload), scaledWorkload)).To(gomega.Succeed())
-					g.Expect(scaledWorkload.Status.Admission).ShouldNot(gomega.BeNil())
-					g.Expect(scaledWorkload.Spec.PodSets).Should(gomega.HaveLen(1))
-					g.Expect(scaledWorkload.Spec.PodSets[0].Count).Should(gomega.Equal(scaledParallelism))
-				}, util.MediumTimeout, util.Interval).Should(gomega.Succeed())
-
-				ta := tas.InternalFrom(scaledWorkload.Status.Admission.PodSetAssignments[0].TopologyAssignment)
-				gomega.Expect(ta).ShouldNot(gomega.BeNil())
-
-				foundOriginal := false
-				for _, domain := range ta.Domains {
-					if len(domain.Values) > 0 && domain.Values[0] == originalNodeName {
-						foundOriginal = true
-						break
-					}
-				}
-				gomega.Expect(foundOriginal).To(gomega.BeTrue(),
-					"original node %s should be preserved in scaled assignment", originalNodeName)
 			})
 		})
 	})
